[mcp] Moved in-memory storage to test project
