+++ src/HotChocolate/ModelContextProtocol/test/HotChocolate.ModelContextProtocol.Tests/IntegrationTests.cs
@@ -26,7 +26,7 @@ public sealed class IntegrationTests
     public async Task ListTools_Valid_ReturnsTools()
     {
         // arrange
-        var storage = new InMemoryOperationToolStorage();
+        var storage = new TestOperationToolStorage();
         await storage.AddOrUpdateToolAsync(
             Utf8GraphQLParser.Parse(
                 await File.ReadAllTextAsync("__resources__/GetWithNullableVariables.graphql")));
@@ -58,7 +58,7 @@ public sealed class IntegrationTests
     public async Task ListTools_AfterSchemaUpdate_ReturnsUpdatedTools()
     {
         // arrange
-        var storage = new InMemoryOperationToolStorage();
+        var storage = new TestOperationToolStorage();
         await storage.AddOrUpdateToolAsync(
             Utf8GraphQLParser.Parse(
                 await File.ReadAllTextAsync("__resources__/GetSingleField.graphql")));
@@ -119,7 +119,7 @@ public sealed class IntegrationTests
     public async Task ListTools_AfterToolsUpdate_ReturnsUpdatedTools()
     {
         // arrange
-        var storage = new InMemoryOperationToolStorage();
+        var storage = new TestOperationToolStorage();
         await storage.AddOrUpdateToolAsync(
             Utf8GraphQLParser.Parse(
                 await File.ReadAllTextAsync("__resources__/GetSingleField.graphql")));
@@ -182,7 +182,7 @@ public sealed class IntegrationTests
     public async Task CallTool_GetWithNullableVariables_ReturnsExpectedResult()
     {
         // arrange
-        var storage = new InMemoryOperationToolStorage();
+        var storage = new TestOperationToolStorage();
         await storage.AddOrUpdateToolAsync(
             Utf8GraphQLParser.Parse(
                 await File.ReadAllTextAsync("__resources__/GetWithNullableVariables.graphql")));
@@ -232,7 +232,7 @@ public sealed class IntegrationTests
     public async Task CallTool_GetWithNonNullableVariables_ReturnsExpectedResult()
     {
         // arrange
-        var storage = new InMemoryOperationToolStorage();
+        var storage = new TestOperationToolStorage();
         await storage.AddOrUpdateToolAsync(
             Utf8GraphQLParser.Parse(
                 await File.ReadAllTextAsync("__resources__/GetWithNonNullableVariables.graphql")));
@@ -283,7 +283,7 @@ public sealed class IntegrationTests
     public async Task CallTool_GetWithDefaultedVariables_ReturnsExpectedResult()
     {
         // arrange
-        var storage = new InMemoryOperationToolStorage();
+        var storage = new TestOperationToolStorage();
         await storage.AddOrUpdateToolAsync(
             Utf8GraphQLParser.Parse(
                 await File.ReadAllTextAsync("__resources__/GetWithDefaultedVariables.graphql")));
@@ -304,7 +304,7 @@ public sealed class IntegrationTests
     public async Task CallTool_GetWithComplexVariables_ReturnsExpectedResult()
     {
         // arrange
-        var storage = new InMemoryOperationToolStorage();
+        var storage = new TestOperationToolStorage();
         await storage.AddOrUpdateToolAsync(
             Utf8GraphQLParser.Parse(
                 await File.ReadAllTextAsync("__resources__/GetWithComplexVariables.graphql")));
@@ -356,7 +356,7 @@ public sealed class IntegrationTests
     public async Task CallTool_GetWithErrors_ReturnsExpectedResult()
     {
         // arrange
-        var storage = new InMemoryOperationToolStorage();
+        var storage = new TestOperationToolStorage();
         await storage.AddOrUpdateToolAsync(
             Utf8GraphQLParser.Parse("query GetWithErrors { withErrors }"));
         var server = CreateTestServer(b => b.AddMcpToolStorage(storage));
@@ -376,7 +376,7 @@ public sealed class IntegrationTests
     public async Task CallTool_GetWithAuthSuccess_ReturnsExpectedResult()
     {
         // arrange
-        var storage = new InMemoryOperationToolStorage();
+        var storage = new TestOperationToolStorage();
         await storage.AddOrUpdateToolAsync(Utf8GraphQLParser.Parse("query GetWithAuth { withAuth }"));
         var server = CreateTestServer(b => b.AddMcpToolStorage(storage));
         var mcpClient = await CreateMcpClientAsync(server.CreateClient(), TestJwtTokenHelper.GenerateToken());
@@ -396,7 +396,7 @@ public sealed class IntegrationTests
     public async Task CallTool_GetWithAuthFailure_ReturnsExpectedResult()
     {
         // arrange
-        var storage = new InMemoryOperationToolStorage();
+        var storage = new TestOperationToolStorage();
         await storage.AddOrUpdateToolAsync(Utf8GraphQLParser.Parse("query GetWithAuth { withAuth }"));
         var server = CreateTestServer(b => b.AddMcpToolStorage(storage));
         var mcpClient = await CreateMcpClientAsync(server.CreateClient());
